services/fastapi/requirements.txt
fastapi
uvicorn[standard]
typedb-client
loguru
python-multipart
